Start work on production builds (#1)
Note that this is about reducing the runtime overhead of dev-mode
constructs, not (yet) about eliminating them from the build.

1. use import.meta.env.PROD in user code
2. add `pnpm test:prod`, which runs the tests in prod mode
3. Make verify() and verified() noops in production
4. Update rollup.config.js to mirror how vite replaces import.meta.env

Note that the reason we have rollup.config.ts is because we need the
feature of rollup that allows us to create separate, distinct configs,
because we don't want the packages to share chunks with each other.
See [this vite issue].

[this vite issue]: https://github.com/vitejs/vite/discussions/1736